Minimize compiling warnings in ltjs_to2_client_shell

+++ game/clientshelldll/to2/screenaudio.cpp
@@ -80,7 +80,7 @@ LTBOOL CScreenAudio::Build()
 	g_pLayoutMgr->GetScreenCustomString(SCREEN_ID_AUDIO,"FrameTexture",szFrame,sizeof(szFrame));
 	HTEXTURE hFrame = g_pInterfaceResMgr->GetTexture(szFrame);
 	CLTGUIFrame *pFrame = debug_new(CLTGUIFrame);
-	pFrame->Create(hFrame,nWd,nHt+8,LTTRUE);
+	pFrame->Create(hFrame,static_cast<uint16>(nWd),static_cast<uint16>(nHt+8),LTTRUE);
 	pFrame->SetBasePos(pos);
 	pFrame->SetBorder(2,m_SelectedColor);
 	AddControl(pFrame);
@@ -180,7 +180,7 @@ void CScreenAudio::OnFocus(LTBOOL bFocus)
 		m_pMusicVolumeCtrl->Enable(m_bMusicEnabled);
 		m_pMusicQualityCtrl->Enable(m_bMusicEnabled && m_nMusicVolume > MUSIC_MIN_VOL);
 
-		m_nPerformance = g_pPerformanceMgr->GetPerformanceCfg(false);
+		m_nPerformance = static_cast<uint8>(g_pPerformanceMgr->GetPerformanceCfg(false));
 
 	    UpdateData(LTFALSE);
 	}
